Refactor HuluApp into a few separate packages instead of one big one
